Updated motor and usart library

+++ libraries&settings/TM_STM32_HAL_LIBRARIES_modified/tm_stm32_usart.h
@@ -238,7 +238,7 @@ typedef enum {
 
 /* NVIC Global Priority */
 #ifndef USART_NVIC_PRIORITY
-#define USART_NVIC_PRIORITY					0x06
+#define USART_NVIC_PRIORITY					((uint32_t)0x06)
 #endif
 
 /* U(S)ART settings, can be changed in your defines.h project file */
